+++ src/domUpdates.js
@@ -34,7 +34,7 @@ export default {
         fancyCategoryNames.push('Cable TV')
       }
     });
-      console.log(fancyCategoryNames);
+      // console.log(fancyCategoryNames);
       $(".category-one").text(fancyCategoryNames[0]);
       $(".category-two").text(fancyCategoryNames[1]);
       $(".category-three").text(fancyCategoryNames[2]);
